requirements.txt
pandas
numpy
holidays
pyyaml
prophet
scikit-learn
matplotlib